+++ src/dune/setup.defaults.ml
@@ -1,3 +1,5 @@
 let library_path = None
 
 let library_destdir = None
+
+let mandir = None